[ C105 ] API data store - Added parameter data_id to requests Set, Get and Del, and response GetList.

+++ enki/handlersrestapi.py
@@ -177,6 +177,7 @@ class HandlerAPIv1DataStoreSet( webapp2.RequestHandler ):
 			user_id = int( jsonobject.get( 'user_id', ''))
 			auth_token = jsonobject.get( 'auth_token', '')
 			data_type = jsonobject.get( 'data_type', '')
+			data_id = jsonobject.get( 'data_id', '')
 			data_payload = jsonobject.get( 'data_payload' )
 			# set the expiry time
 			time_to_expiry = int(jsonobject.get( 'time_expires', enki.librestapi.DATASTORE_EXPIRY_DEFAULT ))  # default lifetime if unspecified: 24h
@@ -184,14 +185,14 @@ class HandlerAPIv1DataStoreSet( webapp2.RequestHandler ):
 				time_to_expiry = enki.librestapi.DATASTORE_NON_EXPIRING   # 100 years
 			time_expires = datetime.datetime.now() + datetime.timedelta( seconds = time_to_expiry )
 			read_access = jsonobject.get( 'read_access', '' )
-			if user_id and auth_token and data_type and data_payload and time_expires:
+			if user_id and auth_token and data_type and data_id and data_payload and time_expires:
 				token_valid = EnkiModelRestAPITokenVerify.get_by_user_id_token( user_id, auth_token )
 				if token_valid:   # user is valid
 					# add optional calculated properties to the data payload
 					if 'calc_ip_addr' in data_payload:    # IP address of the request
 						remote_address = self.request.remote_addr
 						data_payload.update({ 'calc_ip_addr' : remote_address })
-					data_store = enki.librestapi.get_EnkiModelRestAPIDataStore_by_user_id_app_id_data_type( user_id, token_valid.app_id, data_type )
+					data_store = enki.librestapi.get_EnkiModelRestAPIDataStore_by_user_id_app_id_data_type_data_id( user_id, token_valid.app_id, data_type, data_id )
 					if data_store:  # update
 						data_store.data_payload = data_payload
 						data_store.time_expires = time_expires  # update the expiry time
@@ -199,9 +200,9 @@ class HandlerAPIv1DataStoreSet( webapp2.RequestHandler ):
 							data_store.read_access = read_access
 					else:   # create new
 						if not read_access: # if read_access is not specified, use the default value defined in the model
-							data_store = EnkiModelRestAPIDataStore( user_id = user_id, app_id = token_valid.app_id, data_type = data_type, data_payload = data_payload, time_expires = time_expires )
+							data_store = EnkiModelRestAPIDataStore( user_id = user_id, app_id = token_valid.app_id, data_type = data_type, data_id = data_id, data_payload = data_payload, time_expires = time_expires )
 						else:
-							data_store = EnkiModelRestAPIDataStore( user_id = user_id, app_id = token_valid.app_id, data_type = data_type, data_payload = data_payload, time_expires = time_expires, read_access = read_access )
+							data_store = EnkiModelRestAPIDataStore( user_id = user_id, app_id = token_valid.app_id, data_type = data_type, data_id = data_id, data_payload = data_payload, time_expires = time_expires, read_access = read_access )
 					data_store.put()
 					success = True
 					error = ''
@@ -223,10 +224,11 @@ class HandlerAPIv1DataStoreGet( webapp2.RequestHandler ):
 			user_id = int( jsonobject.get( 'user_id', ''))
 			auth_token = jsonobject.get( 'auth_token', '')
 			data_type = jsonobject.get( 'data_type', '')
-			if user_id and auth_token and data_type:
+			data_id = jsonobject.get( 'data_id', '')
+			if user_id and auth_token and data_type and data_id:
 				token_valid = EnkiModelRestAPITokenVerify.get_by_user_id_token( user_id, auth_token )
 				if token_valid:   # user is valid
-					data_store_item = enki.librestapi.get_EnkiModelRestAPIDataStore_by_user_id_app_id_data_type_not_expired( user_id, token_valid.app_id, data_type )
+					data_store_item = enki.librestapi.get_EnkiModelRestAPIDataStore_by_user_id_app_id_data_type_data_id_not_expired( user_id, token_valid.app_id, data_type, data_id )
 					if data_store_item:
 						answer.update({ 'data_payload' : data_store_item.data_payload, 'time_expires' : enki.librestapi.seconds_from_epoch( data_store_item.time_expires ) , 'read_access' : data_store_item.read_access })
 						success = True
@@ -265,7 +267,7 @@ class HandlerAPIv1DataStoreGetList( webapp2.RequestHandler ):
 							for friend_id in friends_list:
 								data_store_item = enki.librestapi.get_EnkiModelRestAPIDataStore_by_user_id_app_id_data_type_read_access_not_expired( friend_id, token_valid.app_id, data_type, read_access )
 								if data_store_item:
-									data_payloads.append({ 'user_id' : str( data_store_item.user_id ), 'data_payload' : data_store_item.data_payload, 'time_expires' : enki.librestapi.seconds_from_epoch( data_store_item.time_expires )})
+									data_payloads.append({ 'user_id' : str( data_store_item.user_id ), 'data_id' : data_store_item.data_id, 'data_payload' : data_store_item.data_payload, 'time_expires' : enki.librestapi.seconds_from_epoch( data_store_item.time_expires )})
 							if data_payloads:
 								answer.update({ 'data_payloads' : data_payloads })
 								success = True
@@ -275,7 +277,7 @@ class HandlerAPIv1DataStoreGetList( webapp2.RequestHandler ):
 						if data_store_list:
 							data_payloads = []
 							for data_store_item in data_store_list:
-								data_payloads.append({ 'user_id' : str( data_store_item.user_id ), 'data_payload' : data_store_item.data_payload, 'time_expires' : enki.librestapi.seconds_from_epoch( data_store_item.time_expires )})
+								data_payloads.append({ 'user_id' : str( data_store_item.user_id ), 'data_id' : data_store_item.data_id, 'data_payload' : data_store_item.data_payload, 'time_expires' : enki.librestapi.seconds_from_epoch( data_store_item.time_expires )})
 							answer.update({ 'data_payloads' : data_payloads })
 							success = True
 							error = ''
@@ -297,10 +299,11 @@ class HandlerAPIv1DataStoreDel( webapp2.RequestHandler ):
 			user_id = int( jsonobject.get( 'user_id', ''))
 			auth_token = jsonobject.get( 'auth_token', '')
 			data_type = jsonobject.get( 'data_type', '')
-			if user_id and auth_token and data_type:
+			data_id = jsonobject.get( 'data_id', '')
+			if user_id and auth_token and data_type and data_id:
 				token_valid = EnkiModelRestAPITokenVerify.get_by_user_id_token( user_id, auth_token )
 				if token_valid:   # user is valid
-					data_stores = enki.librestapi.fetch_EnkiModelRestAPIDataStore_by_user_id_app_id_data_type( user_id, token_valid.app_id, data_type )
+					data_stores = enki.librestapi.fetch_EnkiModelRestAPIDataStore_by_user_id_app_id_data_type_data_id( user_id, token_valid.app_id, data_type, data_id )
 					ndb.delete_multi( data_stores )
 					success = True
 					error = ''
